+++ _build/build.config.sample.php
@@ -1,7 +1,7 @@
 <?php
-/*
+/**
  * FormIt2db/db2FormIt
- * 
+ *
  * Copyright 2013 by Thomas Jakobi <thomas.jakobi@partout.info>
  * 
  * The snippets bases on the code in the following thread in MODX forum