+++ core/components/formit2db/elements/snippets/snippet.db2formit.php
@@ -1,5 +1,5 @@
 <?php
-/*
+/**
  * FormIt2db/db2FormIt
  * 
  * Copyright 2013 by Thomas Jakobi <thomas.jakobi@partout.info>
@@ -24,21 +24,50 @@
  * @package formit2db
  * @subpackage db2formit snippet
  */
-$prefix = $modx->getOption('prefix', $scriptProperties, $modx->getOption(xPDO::OPT_TABLE_PREFIX), true);
-$packagename = $modx->getOption('packagename', $scriptProperties, '', true);
-$classname = $modx->getOption('classname', $scriptProperties, '', true);
-$where = $modx->fromJson($modx->getOption('where', $scriptProperties, '', true));
-$paramname = $modx->getOption('paramname', $scriptProperties, '', true);
-$fieldname = $modx->getOption('fieldname', $scriptProperties, $paramname, true);
-$arrayFormat = $modx->getOption('arrayFormat', $scriptProperties, 'csv', true);
-$arrayFields = $modx->fromJson($modx->getOption('arrayFields', $scriptProperties, '[]', true));
-$ignoreFields = $modx->fromJson($modx->getOption('ignoreFields', $scriptProperties, '[]', true));
-$notFoundRedirect = intval($modx->getOption('notFoundRedirect', $scriptProperties, '0'), true);
+$prefix = $modx->getOption('prefix', $scriptProperties, $modx->getOption(xPDO::OPT_TABLE_PREFIX), TRUE);
+$packagename = $modx->getOption('packagename', $scriptProperties, '', TRUE);
+$classname = $modx->getOption('classname', $scriptProperties, '', TRUE);
+$tablename = $modx->getOption('tablename', $scriptProperties, '', TRUE);
+$where = $modx->fromJson($modx->getOption('where', $scriptProperties, '', TRUE));
+$paramname = $modx->getOption('paramname', $scriptProperties, '', TRUE);
+$fieldname = $modx->getOption('fieldname', $scriptProperties, $paramname, TRUE);
+$arrayFormat = $modx->getOption('arrayFormat', $scriptProperties, 'csv', TRUE);
+$arrayFields = $modx->fromJson($modx->getOption('arrayFields', $scriptProperties, '[]', TRUE));
+$ignoreFields = $modx->fromJson($modx->getOption('ignoreFields', $scriptProperties, '[]', TRUE));
+$notFoundRedirect = (integer) $modx->getOption('notFoundRedirect', $scriptProperties, '0', TRUE);
+$autoPackage = (boolean) $modx->getOption('autoPackage', $scriptProperties, FALSE);
 
 $packagepath = $modx->getOption('core_path') . 'components/' . $packagename . '/';
 $modelpath = $packagepath . 'model/';
 
-$modx->addPackage($packagename, $modelpath, $prefix);
+if ($autoPackage) {
+	$schemapath = $modelpath . 'schema/';
+	$schemafile = $schemapath . $packagename . '.mysql.schema.xml';
+	$manager = $modx->getManager();
+	$generator = $manager->getGenerator();
+	if (!file_exists($schemafile)) {
+
+		if (!is_dir($packagepath)) {
+			mkdir($packagepath, 0777);
+		}
+		if (!is_dir($modelpath)) {
+			mkdir($modelpath, 0777);
+		}
+		if (!is_dir($schemapath)) {
+			mkdir($schemapath, 0777);
+		}
+		//Use this to create a schema from an existing database
+		if (!$generator->writeSchema($schemafile, $packagename, 'xPDOObject', $prefix, TRUE)) {
+			$modx->log(modX::LOG_LEVEL_ERROR, 'Could not generate XML schema in FormIt2db');
+		}
+	}
+	$generator->parseSchema($schemafile, $modelpath);
+	$modx->log(modX::LOG_LEVEL_WARN, 'FormIt2db autoPackage parameter active');
+	$modx->addPackage($packagename, $modelpath, $prefix);
+	$classname = $generator->getClassName($tablename);
+} else {
+	$modx->addPackage($packagename, $modelpath, $prefix);
+}
 
 if ($fieldname) {
 	if (is_array($where)) {
@@ -53,7 +82,7 @@ if (is_array($where)) {
 		if (!is_object($dataobject) || !($dataobject instanceof xPDOObject)) {
 			$errorMsg = 'Failed to create object of type: ' . $classname;
 			$hook->addError('error_message', $errorMsg);
-			return false;
+			return FALSE;
 		}
 		if (empty($dataobject) && $notFoundRedirect) {
 			$modx->sendRedirect($modx->makeUrl($notFoundRedirect));
@@ -66,7 +95,7 @@ if (is_array($where)) {
 			if (in_array($field, $arrayFields)) {
 				switch ($arrayFormat) {
 					case 'json': {
-							$formFields[$field] = json_decode($value, true);
+							$formFields[$field] = json_decode($value, TRUE);
 							break;
 						}
 					case 'csv' :
@@ -85,5 +114,5 @@ if (is_array($where)) {
 	}
 }
 
-return true;
+return TRUE;
 ?>