+++ CHANGELOG.md
@@ -7,3 +7,6 @@ Changelog
     - Add: param ignoreFields to db2Formit
     - Change: paramname/fieldname of db2Formit are compared with the REQUEST array instead of GET array
     - Bugfix: wrong parameter names in db2Formit code (now paramname/fieldname as documented)
+- 1.1
+    - Add: param autoPackage - Autocreate the xPDO Package with packagename and tablename
+    - Add: param tablename - tablename for autocreation of the xPDO Package